wip app refactoring
