+++ app/app.tsx
@@ -3,18 +3,7 @@
 // In this file, we'll be kicking off our app or storybook.
 import "intl-pluralrules"
 import "react-native-reanimated"
-import {
-  ApolloClient,
-  ApolloLink,
-  ApolloProvider,
-  HttpLink,
-  NormalizedCacheObject,
-  split,
-} from "@apollo/client"
-import { getMainDefinition } from "@apollo/client/utilities"
-import { setContext } from "@apollo/client/link/context"
-import { RetryLink } from "@apollo/client/link/retry"
-import AsyncStorage from "@react-native-async-storage/async-storage"
+
 import analytics from "@react-native-firebase/analytics"
 import "@react-native-firebase/crashlytics"
 import {
@@ -23,45 +12,32 @@ import {
   NavigationState,
   PartialState,
 } from "@react-navigation/native"
-import { AsyncStorageWrapper, CachePersistor } from "apollo3-cache-persist"
 import "node-libs-react-native/globals" // needed for Buffer?
 import * as React from "react"
-import { useEffect, useMemo, useRef, useState } from "react"
-import { createNetworkStatusNotifier } from "react-apollo-network-status"
+import { useMemo, useRef, useState } from "react"
 import { Dimensions, Linking } from "react-native"
 import EStyleSheet from "react-native-extended-stylesheet"
 import { RootSiblingParent } from "react-native-root-siblings"
-import VersionNumber from "react-native-version-number"
 import { GlobalErrorToast } from "./components/global-error"
-import { createCache } from "./graphql/cache"
 import "./utils/polyfill"
 import { RootStack } from "./navigation/root-navigator"
-import { isIos } from "./utils/helper"
-import { saveString, loadString } from "./utils/storage"
-import useToken, { getAuthorizationHeader } from "./hooks/use-token"
 import ErrorBoundary from "react-native-error-boundary"
 import { ErrorScreen } from "./screens/error-screen"
 // import moment locale files so we can display dates in the user's language
 import "moment/locale/es"
 import "moment/locale/fr-ca"
 import "moment/locale/pt-br"
-import { PriceContextProvider } from "./store/price-context"
 import { AuthenticationContext } from "./store/authentication-context"
 import { LocalizationContextProvider } from "./store/localization-context"
 import { loadAllLocales } from "./i18n/i18n-util.sync"
 import TypesafeI18n from "./i18n/i18n-react"
 import { customLocaleDetector } from "./utils/locale-detector"
-import { useAppConfig } from "./hooks"
 import { ThemeProvider } from "@rneui/themed"
 import theme from "./rne-theme/theme"
-import { createClient } from "graphql-ws"
-import { GraphQLWsLink } from "@apollo/client/link/subscriptions"
 import { GaloyToast } from "./components/galoy-toast"
-import { BUILD_VERSION } from "@app/config"
 import { RootStackParamList } from "./navigation/stack-param-lists"
-
-export const { link: linkNetworkStatusNotifier, useApolloNetworkStatus } =
-  createNetworkStatusNotifier()
+import useToken from "./hooks/use-token"
+import { GaloyClient } from "./graphql/client"
 
 const entireScreenWidth = Dimensions.get("window").width
 EStyleSheet.build({
@@ -69,33 +45,14 @@ EStyleSheet.build({
   // $textColor: '#0275d8'
 })
 
-const noRetryOperations = [
-  "intraLedgerPaymentSend",
-  "intraLedgerUsdPaymentSend",
-
-  "lnInvoiceFeeProbe",
-  "lnInvoicePaymentSend",
-  "lnNoAmountInvoiceFeeProbe",
-  "lnNoAmountInvoicePaymentSend",
-  "lnNoAmountUsdInvoiceFeeProbe",
-  "lnUsdInvoiceFeeProbe",
-  "lnNoAmountUsdInvoicePaymentSend",
-
-  "onChainPaymentSend",
-  "onChainTxFee",
-]
-
 loadAllLocales()
 
 /**
  * This is the root component of our app.
  */
 export const App = (): JSX.Element => {
-  const { token, hasToken, saveToken } = useToken()
-  const { appConfig } = useAppConfig()
+  const { hasToken } = useToken()
   const routeName = useRef("Initial")
-  const [apolloClient, setApolloClient] = useState<ApolloClient<NormalizedCacheObject>>()
-  const [persistor, setPersistor] = useState<CachePersistor<NormalizedCacheObject>>()
   const [isAppLocked, setIsAppLocked] = useState(true)
   const processLink = useRef<((url: string) => void) | null>(null)
   processLink.current = () => {
@@ -128,130 +85,6 @@ export const App = (): JSX.Element => {
     }
 
     return route.name
-  }
-
-  useEffect(() => {
-    const fn = async () => {
-      const httpLink = new HttpLink({
-        uri: appConfig.galoyInstance.graphqlUri,
-      })
-
-      // TODO: used to migrate from jwt to kratos token, remove after a few releases
-      const updateTokenLink = new ApolloLink((operation, forward) => {
-        return forward(operation).map((response) => {
-          const context = operation.getContext()
-
-          const kratosToken = context.response.headers.get("kratos-session-token")
-
-          if (kratosToken) {
-            saveToken(kratosToken)
-          }
-
-          return response
-        })
-      })
-
-      const wsLink = new GraphQLWsLink(
-        createClient({
-          url: appConfig.galoyInstance.graphqlWsUri,
-          connectionParams: hasToken ? { Authorization: `Bearer ${token}` } : undefined,
-          // Voluntary not using: webSocketImpl: WebSocket
-          // seems react native already have an implement of the websocket?
-          //
-          // TODO: implement keepAlive and reconnection?
-          // https://github.com/enisdenjo/graphql-ws/blob/master/docs/interfaces/client.ClientOptions.md#keepalive
-        }),
-      )
-
-      const splitLink = split(
-        ({ query }) => {
-          const definition = getMainDefinition(query)
-          return (
-            definition.kind === "OperationDefinition" &&
-            definition.operation === "subscription"
-          )
-        },
-        wsLink,
-        updateTokenLink.concat(httpLink),
-      )
-
-      const authLink = setContext((request, { headers }) => {
-        return {
-          headers: {
-            ...headers,
-            authorization: token ? getAuthorizationHeader(token) : "",
-          },
-        }
-      })
-
-      const retryLink = new RetryLink({
-        delay: {
-          initial: 500, // default = 300
-          // max: Infinity,
-          // jitter: true
-        },
-        attempts: {
-          max: 3,
-          retryIf: (error, operation) => {
-            console.debug(JSON.stringify(error), "retry error test")
-            return Boolean(error) && !noRetryOperations.includes(operation.operationName)
-          },
-        },
-      })
-
-      const cache = createCache()
-
-      const persistor_ = new CachePersistor({
-        cache,
-        storage: new AsyncStorageWrapper(AsyncStorage),
-        debug: __DEV__,
-      })
-
-      setPersistor(persistor_)
-
-      const client = new ApolloClient({
-        cache,
-        link: linkNetworkStatusNotifier.concat(
-          retryLink.concat(authLink.concat(splitLink)),
-        ),
-        name: isIos ? "iOS" : "Android",
-        version: `${VersionNumber.appVersion}-${VersionNumber.buildVersion}`,
-        connectToDevTools: true,
-      })
-
-      // Read the current schema version from AsyncStorage.
-      const currentVersion = await loadString(BUILD_VERSION)
-      const buildVersion = String(VersionNumber.buildVersion)
-
-      // TODO: also add a schema version?
-      if (currentVersion === buildVersion) {
-        // If the current version matches the latest version,
-        // we're good to go and can restore the cache.
-        await persistor_.restore()
-      } else {
-        // Otherwise, we'll want to purge the outdated persisted cache
-        // and mark ourselves as having updated to the latest version.
-
-        // init the DB. will be override if a cache exists
-        await persistor_.purge()
-        await saveString(BUILD_VERSION, buildVersion)
-      }
-
-      setApolloClient(client)
-    }
-    fn()
-  }, [appConfig.galoyInstance, token, hasToken, saveToken])
-
-  // Before we show the app, we have to wait for our state to be ready.
-  // In the meantime, don't render anything. This will be the background
-  // color set in native by rootView's background color.
-  //
-  // This step should be completely covered over by the splash screen though.
-  //
-  // You're welcome to swap in your own component to render if your boot up
-  // sequence is too slow though.
-  if (!apolloClient || !persistor) {
-    return <></>
   }
 
   const linking: LinkingOptions<RootStackParamList> = {
@@ -295,38 +128,36 @@ export const App = (): JSX.Element => {
   return (
     <AuthenticationContext.Provider value={{ isAppLocked, setAppUnlocked, setAppLocked }}>
       <ThemeProvider theme={theme}>
-        <ApolloProvider client={apolloClient}>
-          <PriceContextProvider>
-            <TypesafeI18n locale={customLocaleDetector()}>
-              <LocalizationContextProvider>
-                <ErrorBoundary FallbackComponent={ErrorScreen}>
-                  <NavigationContainer
-                    linking={linking}
-                    onStateChange={(state) => {
-                      const currentRouteName = getActiveRouteName(state)
+        <GaloyClient>
+          <TypesafeI18n locale={customLocaleDetector()}>
+            <LocalizationContextProvider>
+              <ErrorBoundary FallbackComponent={ErrorScreen}>
+                <NavigationContainer
+                  linking={linking}
+                  onStateChange={(state) => {
+                    const currentRouteName = getActiveRouteName(state)
 
-                      if (routeName.current !== currentRouteName && currentRouteName) {
-                        /* eslint-disable camelcase */
-                        analytics().logScreenView({
-                          screen_name: currentRouteName,
-                          screen_class: currentRouteName,
-                          is_manual_log: true,
-                        })
-                        routeName.current = currentRouteName
-                      }
-                    }}
-                  >
-                    <RootSiblingParent>
-                      <GlobalErrorToast />
-                      <RootStack />
-                      <GaloyToast />
-                    </RootSiblingParent>
-                  </NavigationContainer>
-                </ErrorBoundary>
-              </LocalizationContextProvider>
-            </TypesafeI18n>
-          </PriceContextProvider>
-        </ApolloProvider>
+                    if (routeName.current !== currentRouteName && currentRouteName) {
+                      /* eslint-disable camelcase */
+                      analytics().logScreenView({
+                        screen_name: currentRouteName,
+                        screen_class: currentRouteName,
+                        is_manual_log: true,
+                      })
+                      routeName.current = currentRouteName
+                    }
+                  }}
+                >
+                  <RootSiblingParent>
+                    <GlobalErrorToast />
+                    <RootStack />
+                    <GaloyToast />
+                  </RootSiblingParent>
+                </NavigationContainer>
+              </ErrorBoundary>
+            </LocalizationContextProvider>
+          </TypesafeI18n>
+        </GaloyClient>
       </ThemeProvider>
     </AuthenticationContext.Provider>
   )
